Changed behavior of the Levels image effect. I also renamed it -- High-Mid-Low. This more accurately describes the conventional terms for this operation.
I unlocked the thresholds for the high and low points. This was from on-set feedback working with a VFX technician -- they wanted the ability to play a larger range.

Unlocking the thresholds does allow some irregular mid behavior when you have your high and low points inverted, but that's not a terrible thing.